Fix app not building when adding components dir
The limb_ble_periph component uses NimBLE, which isn't enabled by
default and can't be force enabled by the component itself (at least I
haven't found a way to do it. Since it's enabled using a `choice` in
Kconfig, it's apparently not possible to use a `select` to force enable
it). If you don't do a minimal build, every component in the components
directory gets built whenever you build your own app, regardless if you
actually use the component. If you then don't have the NimBLE component
enabled in your app, the build will fail when trying to build the
limb_ble_periph component.

The simple solution is to just always do a minimal build so we don't
build components if we don't use them.