Don't use a loop for message data in can_receive
Since we don't support CAN FD, the max size of message data is 8 bytes.
That means we can do a direct copy treating the buffer as an 8-byte int
instead of using a loop. Doing it this way we only need two loads and
two stores for the copy in an optimized build:
(https://godbolt.org/z/h6EbKjfrT).

`can_send` is more of an issue since the user might pass a pointer to a
buffer that's smaller than 8 bytes. We could get around this by creating
an argument struct containing an 8-byte buffer that we take as an
argument by pointer. That way it's always safe to read 8 bytes.