adc_manager: Handle timeouts properly in `adc_mgr_read`
Since the ADC manager supports using a lower sample rate than the
actual sample rate that the ADC peripheral uses, sometimes we end up
reading values that aren't useful to the caller, and as such they are
discarded. If we then in a subsequent call to `adc_continuous_read` get
a timeout, we might not have written any actual useful value, so in
that case we have actually had a proper timeout from our perspective,
even though we ended up reading some values from the ADC.

Additionally, the `timeout_ms` value passed by the caller should apply
to the overall process inside `adc_mgr_read`, so we shouldn't pass it
directly to `adc_continuous_read`. If we do, every new call to
`adc_continuous_read` will wait for the full timeout period, even if
previous calls to it already waited for some time.

Actually return an `ESP_ERR_TIMEOUT` if