servo: Rework component for proper continuous movement
The initial attempt at connecting our own potentiometer to the servo
didn't work due to unstable voltages when the motor switched between
running and being still. Our solution is to remove the servo's
potentiometer entirely, making it so that the servo thinks its
potentiometer is always at the midway point, which corresponds to a
pulse width of 1500 us. By setting the pulse width to values above and
below 1500 us, we end up being able to roughly control the angular
velocity of the servo. This allows us to control its movement using the
`servo_update` function.